Refactor symbol annotations into modular package structure
Extract monolithic _annotations.py into focused modules within new
_symbol_annotations package for better code organization and maintainability.

- Extract 963 lines from dagster/_annotations.py
- Create _symbol_annotations/ package with 4 focused modules:
  - annotatable.py: Core annotatable functionality
  - lifecycle.py: Lifecycle management system
  - public.py: Public API annotation handling
  - __init__.py: Package exports and initialization
