+++ python_modules/dagster/dagster/_symbol_annotations/__init__.py
@@ -1,0 +1,73 @@
+# Re-export all public symbols for backward compatibility
+from dagster._symbol_annotations.annotatable import Annotatable, T_Annotatable, copy_annotations
+from dagster._symbol_annotations.lifecycle import (
+    BetaInfo,
+    DeprecatedInfo,
+    PreviewInfo,
+    SupersededInfo,
+    attach_deprecation_info_and_wrap,
+    beta,
+    beta_param,
+    deprecated,
+    deprecated_param,
+    get_beta_info,
+    get_beta_param_info,
+    get_beta_params,
+    get_deprecated_info,
+    get_deprecated_param_info,
+    get_deprecated_params,
+    get_preview_info,
+    get_superseded_info,
+    has_beta_params,
+    has_deprecated_params,
+    hidden_param,
+    is_beta,
+    is_beta_param,
+    is_deprecated,
+    is_deprecated_param,
+    is_preview,
+    is_superseded,
+    only_allow_hidden_params_in_kwargs,
+    preview,
+    superseded,
+)
+from dagster._symbol_annotations.public import PUBLIC, PublicAttr, is_public, public
+
+__all__ = [
+    "PUBLIC",
+    "Annotatable",
+    "BetaInfo",
+    "DeprecatedInfo",
+    "PreviewInfo",
+    "PublicAttr",
+    "SupersededInfo",
+    "T_Annotatable",
+    "attach_deprecation_info_and_wrap",
+    "beta",
+    "beta_param",
+    "copy_annotations",
+    "deprecated",
+    "deprecated_param",
+    "get_beta_info",
+    "get_beta_param_info",
+    "get_beta_params",
+    "get_deprecated_info",
+    "get_deprecated_param_info",
+    "get_deprecated_params",
+    "get_preview_info",
+    "get_superseded_info",
+    "has_beta_params",
+    "has_deprecated_params",
+    "hidden_param",
+    "is_beta",
+    "is_beta_param",
+    "is_deprecated",
+    "is_deprecated_param",
+    "is_preview",
+    "is_public",
+    "is_superseded",
+    "only_allow_hidden_params_in_kwargs",
+    "preview",
+    "public",
+    "superseded",
+]
